raplcap-ipg: new implementation using libraplcap-ipg
The Intel Power Gadget API version used by libraplcap-ipg appears to be
deprecated in IPG version 3.7 and doesn't work properly on macOS.

EnergyMon now has an "ipg" implementation that uses Intel Power Gadget
directly, which is more appropriate since RAPLCap can't support its complete
API with IPG as the backend anyway (a backend which is also not officially
supported in RAPLCap).

This branch is being created so we have a record of this code in case it proves
useful in the future.